requirements.txt
<<<<<<< HEAD
backports-abc==0.5
bkcharts==0.2
blinker==1.4
bokeh==0.12.6
=======
blinker==1.4
>>>>>>> master
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Login==0.4.0
<<<<<<< HEAD
Flask-Principal==0.4.0
=======
Flask-Mail==0.9.1
flask-principal==0.4.0
>>>>>>> master
Flask-SQLAlchemy==2.2
futures==3.1.1
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
<<<<<<< HEAD
Jinja2==2.9.6
MarkupSafe==1.0
MySQL-python==1.2.5
numpy==1.13.1
pandas==0.20.3
=======
>>>>>>> master
PyGithub==1.34
PyJWT==1.5.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.1
<<<<<<< HEAD
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.11
tornado==4.5.1
=======
>>>>>>> master
urllib3==1.21.1
Werkzeug==0.12.2
